fix(protocol): widen frame header frameOff to u32 (#46)
The frameOff field encodes a byte offset within the block (not the file).
At the default 1 MiB block framed at 16 KiB, offsets reach ~1 MiB, far
past u16's 64 KiB limit. Offset 65536 wraps to 0, falsely signaling a new
block mid-block and corrupting reassembly ("short block" FailIntegrity).

Widen frameOff from u16 to u32. The header stays 16 bytes by consuming
the former trailing reserved(u16). New layout (big-endian):
  version(u8) type(u8) flags(u8) reserved(u8)
  fileIdx(u16) blockIdx(u32) frameOff(u32) len(u16)

Updated in lockstep: TS encoder/decoder, Go wire frame + AEAD + sender +
receiver, and the shared AEAD test vector (GCM tag changed since AAD
changed; ciphertext unchanged).

+++ packages/wire/transfer_receiver.go
@@ -147,7 +147,7 @@ func (r *Receiver) applyManifest(payload []byte) error {
 	return nil
 }
 
-func (r *Receiver) onBlockData(blockIdx uint32, frameOff uint16, payload []byte) error {
+func (r *Receiver) onBlockData(blockIdx uint32, frameOff uint32, payload []byte) error {
 	if r.file == nil {
 		return NewTransferError(FailIntegrity, "block_data before manifest")
 	}
